REVERT migration to float

+++ authenticate-me/backend/db/migrations/20230414221800-create-review.js
@@ -33,7 +33,7 @@ module.exports = {
         allowNull: false,
       },
       stars: {
-        type: Sequelize.DECIMAL(3,2),// min 1 max 5
+        type: Sequelize.FLOAT,// min 1 max 5
         allowNull: false,
       },
       createdAt: {
